+++ CHANGES.md
@@ -1,6 +1,12 @@
+dev
+------
+
 - Permute the order of Yum groupinstall/install to workaround
-  a build issue in CentOS 8 under OverlayFS/Docker.
-- Do not install `yum-ovl-plugin` workaround on CentOS 8.
+  a build issue in CentOS 8 under OverlayFS/Docker. (@avsm)
+- Do not install `yum-ovl-plugin` workaround on CentOS 8. (@avsm)
+- Add Fedora 31 and Ubuntu 19.10 (@XVilka @talex5)
+- Add Alpine 3.11 and Ubuntu 20.04 (@avsm)
+- Remove Ubuntu 19.04 from the supported distro list (@avsm).
 
 v6.3.0 2019-08-11 Cambridge
 ---------------------------